simulation/backend/requirements.txt
fastapi
huggingface_hub
joblib
pandas
python-dotenv
scikit-learn
uvicorn[standard]

